requirements.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Model Persistence
joblib>=1.3.0

# Data Version Control
dvc>=3.0.0

# Configuration Management
pyyaml>=6.0

# FastAPI and Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# API Testing
requests>=2.31.0

# Optional: For better performance
python-multipart>=0.0.6
